+++ src/main/java/com/inz/airline/domain/Ticket.java
@@ -1,6 +1,5 @@
 package com.inz.airline.domain;
 
-import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.neo4j.ogm.annotation.GeneratedValue;
@@ -8,7 +7,6 @@ import org.neo4j.ogm.annotation.Id;
 import org.neo4j.ogm.annotation.NodeEntity;
 
 @NoArgsConstructor
-@AllArgsConstructor
 @Data
 @NodeEntity
 public class Ticket {
