+++ src/main/java/com/inz/airline/domain/UserRequest.java
@@ -12,7 +12,6 @@ public class UserRequest {
   private String password;
   private String firstname;
   private String lastname;
-  private String email;
   private String nationality;
   private LocalDate dateOfBirth;
 }
